server/requirements.txt
fastapi==0.109.2
uvicorn==0.27.1
groq==0.7.1
python-dotenv==1.0.1
pydantic==2.6.4
SpeechRecognition==3.10.1
PyAudio==0.2.14
pyttsx3==2.90
